Move the errors into their own module.
